Moved padding based on border radius into its own file
